+++ packages/elements/emoji/src/__tests__/unit/api/_resource-spec-util.ts
@@ -1,6 +1,4 @@
-import { ServiceConfig } from '@atlaskit/util-service-support';
 import { EmojiResource } from '../../../api/EmojiResource';
-import { EmojiResponse } from '../../../types';
 
 /**
  * #initSiteEmojiResource throws error
@@ -10,10 +8,7 @@ export class ErrorEmojiResource extends EmojiResource {
     return this.activeLoaders;
   }
 
-  initSiteEmojiResource(
-    emojiResponse: EmojiResponse,
-    provider: ServiceConfig,
-  ): Promise<void> {
+  initSiteEmojiResource(): Promise<void> {
     throw new Error('Cannot initSiteEmojiResource');
   }
 }